# deliverables/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Sentinel AML \u2014 Transaction Monitoring System",
        "description": "Real-time money-laundering detection for MeridianTrust Bank.\n\n**Flow:** ingestion \u2192 detection \u2192 risk-scored alert \u2192 case disposition.\n\n**Roles** (RBAC is enforced at this API layer, not only in the UI):\n- `ANALYST` \u2014 alert queue (PII masked), cases, dispositions\n- `SENIOR_ANALYST` \u2014 the above, plus unmasked PII and SAR escalation\n- `COMPLIANCE_ADMIN` \u2014 the above, plus ingestion and rule configuration\n\nDetection thresholds are tunable at runtime via\n`PATCH /api/v1/admin/rules/{ruleCode}` \u2014 no redeployment required.\n",
        "contact": {
            "name": "MeridianTrust Financial Crime Engineering"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "security": [
        {
            "basicAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/cases/{caseRef}/status": {
            "patch": {
                "tags": [
                    "Cases"
                ],
                "summary": "Transition a case",
                "description": "Closing a case requires a disposition and reason, and cascades that\ndisposition onto every member alert \u2014 each cascade separately audited.\n",
                "operationId": "transition",
                "parameters": [
                    {
                        "name": "caseRef",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Transition applied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing disposition or reason",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role does not permit this disposition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Illegal transition for the current state",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Detail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/alerts/{alertRef}/status": {
            "patch": {
                "tags": [
                    "Alerts"
                ],
                "summary": "Transition an alert",
                "description": "Moves an alert through its lifecycle. Closing requires a disposition\nand a reason, both retained permanently with the analyst's identity \u2014\nalerts are never deleted (business rule 6).\n\nESCALATED_TO_SAR requires the SENIOR_ANALYST role.\n",
                "operationId": "transition_1",
                "parameters": [
                    {
                        "name": "alertRef",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AlertTransitionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Transition applied",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing disposition or reason",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Role does not permit this disposition",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Concurrently modified \u2014 reload and retry",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Illegal transition for the current state",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertDetailDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/customers/{customerId}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Customer (PII masked)",
                "description": "Business rule 8: names and identifiers are masked. Any analyst may call this.",
                "operationId": "masked",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Masked"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customerId}/timeline": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Transaction timeline",
                "description": "Chronological activity, annotating each transaction with the alerts it evidences.",
                "operationId": "timeline",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TimelineEntry"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/customers/{customerId}/full": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Customer with full PII",
                "description": "Business rule 8: unmasked KYC detail, restricted to SENIOR_ANALYST\nand above. Enforced by both the URL rule and a method-level check \u2014\nthe UI is not part of the decision.\n",
                "operationId": "full",
                "parameters": [
                    {
                        "name": "customerId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full record returned",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Full"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Requires SENIOR_ANALYST",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Full"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such customer",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Full"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/audit": {
            "get": {
                "tags": [
                    "Audit"
                ],
                "summary": "Audit trail",
                "description": "Every alert and case state transition with its actor and timestamp.\nFilter by entityType (ALERT, CASE, RULE, REFERENCE, INGESTION) and\noptionally by a specific entityId.\n",
                "operationId": "trail",
                "parameters": [
                    {
                        "name": "entityType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "entityId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseAuditEntryDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuditEntry": {
                "type": "object",
                "description": "One immutable audit record.",
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "actor": {
                        "type": "string"
                    },
                    "fromState": {
                        "type": "string"
                    },
                    "toState": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "occurredAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "Detail": {
                "type": "object",
                "description": "Full case view including member alerts and audit trail.",
                "properties": {
                    "caseRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "customerNameMasked": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "priority": {
                        "type": "string"
                    },
                    "aggregateRiskScore": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "assignedTo": {
                        "type": "string"
                    },
                    "narrative": {
                        "type": "string"
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "dispositionReason": {
                        "type": "string"
                    },
                    "openedBy": {
                        "type": "string"
                    },
                    "openedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "closedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "alerts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AlertSummaryDto"
                        }
                    },
                    "auditTrail": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditEntry"
                        }
                    }
                }
            },
...
            "TransitionRequest": {
                "type": "object",
                "description": "Case state transition. Closing requires a disposition and reason.",
                "properties": {
                    "targetStatus": {
                        "type": "string",
                        "enum": [
                            "NEW",
                            "INVESTIGATING",
                            "PENDING_APPROVAL",
                            "CLOSED"
                        ]
                    },
                    "disposition": {
                        "type": "string",
                        "enum": [
                            "TRUE_POSITIVE",
                            "FALSE_POSITIVE",
                            "ESCALATED_TO_SAR",
                            "CLEARED_NO_ACTION"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "maxLength": 4000,
                        "minLength": 0
                    }
                },
                "required": [
                    "targetStatus"
                ]
            },
...
            "AlertTransitionRequest": {
                "type": "object",
                "description": "Alert status transition. Closing requires a disposition and reason (business rule 6).",
                "properties": {
                    "targetStatus": {
                        "type": "string",
                        "enum": [
                            "OPEN",
                            "IN_REVIEW",
                            "ESCALATED",
                            "CLOSED"
                        ]
                    },
                    "disposition": {
                        "type": "string",
                        "description": "Required when targetStatus is CLOSED",
                        "enum": [
                            "TRUE_POSITIVE",
                            "FALSE_POSITIVE",
                            "ESCALATED_TO_SAR",
                            "CLEARED_NO_ACTION"
                        ]
                    },
                    "reason": {
                        "type": "string",
                        "description": "Required when targetStatus is CLOSED",
                        "example": "Verified as salary payments",
                        "maxLength": 2000,
                        "minLength": 0
                    }
                },
                "required": [
                    "targetStatus"
                ]
            },
            "AlertDetailDto": {
                "type": "object",
                "description": "Alert detail with evidence transactions and risk score breakdown.",
                "properties": {
                    "alertRef": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "string"
                    },
                    "customerName": {
                        "type": "string"
                    },
                    "accountId": {
                        "type": "string"
                    },
                    "ruleCode": {
                        "type": "string"
                    },
                    "typology": {
                        "type": "string"
                    },
                    "riskScore": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "severity": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "explanation": {
                        "type": "string",
                        "description": "Human-readable justification generated from the evidence"
                    },
                    "scoreBreakdown": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "description": "How the risk score was composed"
                    },
                    "evidence": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EvidenceTransactionDto"
                        }
                    },
                    "evidenceAmountBase": {
                        "type": "number"
                    },
                    "triggerCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "disposition": {
                        "type": "string"
                    },
                    "dispositionReason": {
                        "type": "string"
                    },
                    "disposedBy": {
                        "type": "string",
                        "description": "Analyst who disposed the alert (business rule 6)"
                    },
                    "disposedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "caseRef": {
                        "type": "string"
                    },
                    "firstDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastDetectedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Masked": {
                "type": "object",
                "description": "Customer with PII masked \u2014 safe for list views and any authenticated analyst.",
                "properties": {
                    "customerId": {
                        "type": "string"
                    },
                    "nameMasked": {
                        "type": "string",
                        "example": "K****** S*****"
                    },
                    "nationalIdMasked": {
                        "type": "string",
                        "example": "*****4821"
                    },
                    "emailMasked": {
                        "type": "string",
                        "example": "k******@g****.com"
                    },
                    "phoneMasked": {
                        "type": "string",
                        "example": "+91-******2955"
                    },
                    "city": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "customerSegment": {
                        "type": "string"
                    },
                    "kycStatus": {
                        "type": "string"
                    },
                    "riskRating": {
                        "type": "string"
                    },
                    "politicallyExposed": {
                        "type": "boolean"
                    }
                }
            },
            "TimelineEntry": {
                "type": "object",
                "description": "Chronological transaction activity for the customer timeline view.",
                "properties": {
                    "transactionId": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "accountId": {
                        "type": "string"
                    },
                    "direction": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "number"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "amountBase": {
                        "type": "number"
                    },
                    "channel": {
                        "type": "string"
                    },
                    "counterpartyName": {
                        "type": "string"
                    },
                    "counterpartyCountry": {
                        "type": "string"
                    },
                    "alertRefs": {
                        "type": "array",
                        "description": "Alert references this transaction is evidence for",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Full": {
                "type": "object",
                "description": "Full customer KYC detail. Requires SENIOR_ANALYST (business rule 8).",
                "properties": {
                    "customerId": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "nationalId": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "occupation": {
                        "type": "string"
                    },
                    "annualIncome": {
                        "type": "number"
                    },
                    "employmentStatus": {
                        "type": "string"
                    },
                    "customerSince": {
                        "type": "string",
                        "format": "date"
                    },
                    "customerSegment": {
                        "type": "string"
                    },
                    "kycStatus": {
                        "type": "string"
                    },
                    "riskRating": {
                        "type": "string"
                    },
                    "politicallyExposed": {
                        "type": "boolean"
                    },
                    "numComplaintsLastYear": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "PageResponseAuditEntryDto": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AuditEntryDto"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "description": "HTTP Basic. Demo users: analyst / senior / admin.",
                "scheme": "basic"
            }
        }
    }
}